move rmtree function into existing module, avoid import as, remove obsolete imports
